requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
imbalanced-learn>=0.8.0
fancyimpute>=0.7.0
openpyxl>=3.0.0

matplotlib>=3.4.0
seaborn>=0.11.0
scipy>=1.7.0
statsmodels>=0.12.0
xgboost>=1.4.0
lightgbm>=3.2.0
catboost>=0.26.0
tensorflow>=2.6.0
keras>=2.6.0
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.11.0
nltk>=3.6.0
spacy>=3.1.0
flask>=2.0.0
flask-cors>=3.0.10
requests>=2.25.0
beautifulsoup4>=4.9.0
pytest>=6.2.0   
pytest-cov>=2.12.0
black>=21.7b0
pre-commit>=2.14.0
jupyter>=1.0.0
jupyterlab>=3.0.0
notebook>=6.4.0
ipython>=7.26.0
ipykernel>=6.5.0    
pydantic>=1.8.0
sqlalchemy>=1.4.0
sqlalchemy-utils>=0.37.8
pymysql>=1.0.2
psycopg2-binary>=2.9.1
redis>=3.5.3
boto3>=1.18.0
azure-storage-blob>=12.8.0
google-cloud-storage>=1.42.0
pyarrow>=4.0.0
fastparquet>=0.7.1
pyodbc>=4.0.30
pyyaml>=5.4.0
pycryptodome>=3.10.1
joblib>=1.0.1
# Additional libraries for specific tasks
sqlalchemy-mixins>=0.9.0
sqlalchemy-json>=0.3.0
sqlalchemy-utils>=0.37.8
sqlalchemy-migrate>=0.13.0
sqlalchemy-databases>=0.1.0
sqlalchemy-utc>=0.2.0
sqlalchemy-jsonapi>=0.1.0
sqlalchemy-audit>=0.1.0
sqlalchemy-continents>=0.1.0
sqlalchemy-geojson>=0.1.0
matplotlib.pyplot>=3.4.0
sqlalchemy-serializer>=0.1.0
seaborn>=0.11.0
statsmodels>=0.12.0